pyproject.toml
[tool.poetry]
name = "research-assistant"
version = "0.1.0"
description = ""
authors = ["Vishwas Gowda <vishwas.g217@gmail.com>"]
readme = "README.md"
packages = [{ include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.51.0"
beautifulsoup4 = "^4.12.3"
cohere = "^5.11.0"
python-dotenv = "^1.0.1"
langchain-text-splitters = "^0.3.0"
ipykernel = "^6.29.5"
weaviate-client = "^4.8.1"
streamlit = "^1.39.0"
watchdog = "^5.0.3"

[tool.pytest.ini_options]
pythonpath = "src"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
